requirements.txt
# Core Framework
fastapi==0.115.5
uvicorn==0.32.1
httpx==0.27.2
pydantic==2.9.2
python-dotenv==1.0.1
streamlit==1.40.1

# LangGraph & Agents
langgraph==0.2.37
langchain-core==0.3.15

# Vector Store & Embeddings
chromadb==0.5.11
sentence-transformers==3.2.1

# LLM Providers
#openai==1.54.3
# anthropic==0.39.0  
google-generativeai==0.3.0  

# Text Processing
rapidfuzz==3.9.7
pandas==2.2.2

# ASR (Whisper)
openai-whisper==20240930
# Requires ffmpeg: brew install ffmpeg (macOS) or apt-get install ffmpeg (Linux)

# Data Formats
pyarrow==17.0.0

# Web Search
# brave-search-python==1.0.0  # If using Brave SDK (optional)

# Utilities
tqdm==4.66.1
